feat: Complete OpenAPI integration for setup and health endpoints
- Implement OpenAPI documentation for both /api/v1/health and /api/v1/setup
- Use @hono/zod-openapi for auto-generated documentation
- Both GET and POST methods for setup endpoint with comprehensive schemas
- Direct mounting to OpenAPIHono for proper documentation generation
- Fix test compatibility with OpenAPI validation (BAD_REQUEST vs INVALID_JSON)
- Maintain consistent file structure: route.ts for OpenAPI schemas
- All tests passing (51/51 for setup endpoints)
- Swagger UI available at /api/v1/ui with complete endpoint documentation
