+++ src/api/index.ts
@@ -4,7 +4,7 @@ import { cors } from 'hono/cors';
 import { swaggerUI } from '@hono/swagger-ui';
 import { healthRoute } from './v1/health/route';
 import { healthHandler } from './v1/health/get';
-import { setupStatusRoute, setupConfigRoute } from './v1/setup/openapi';
+import { setupStatusRoute, setupConfigRoute } from './v1/setup/route';
 import { setupGetHandler } from './v1/setup/get';
 import { setupPostHandler } from './v1/setup/post';
 import sheetsRouter from './v1/sheets/route';
@@ -40,8 +40,6 @@ api.openapi(healthRoute, healthHandler);
 // Setup routes - direct OpenAPI mount
 api.openapi(setupStatusRoute, setupGetHandler);
 api.openapi(setupConfigRoute, setupPostHandler);
-
-// Setup routes now handled by direct OpenAPI mount above
 
 // Sheets routes
 v1.route('/sheets', sheetsRouter);
